+++ tests/testthat/test-create_codemeta.R
@@ -1,14 +1,5 @@
-testthat::test_that("we can create a codemeta document given a package name", {
-  skip_on_cran()
-  skip_if_offline()
-  skip_on_ci()
-  testthat::expect_silent(
-    create_codemeta("utils", path = path, verbose = FALSE)
-  )
 
-})
-
-testthat::test_that("We can deduce relatedLink from installed pkg", {
+test_that("We can deduce relatedLink from installed pkg", {
   skip_on_cran()
   skip_if_offline()
   skip_on_ci()
@@ -19,10 +10,3 @@ testthat::test_that("We can deduce relatedLink from installed pkg", {
 
 })
 
-## Test that we can write codemeta from a temp working dir (e.g. non-root dir)
-testthat::test_that("we can write codemeta given a codemeta object", {
-  skip_on_cran()
-  skip_if_offline()
-  codemeta <- new_codemeta()
-  expect_type(create_codemeta("codemetar", codemeta, verbose = FALSE), "list")
-})